+++ base/src/main/java/gamelauncher/engine/render/GameItem.java
@@ -8,8 +8,8 @@ import gamelauncher.engine.resource.AbstractGameResource;
 import gamelauncher.engine.util.GameException;
 import gamelauncher.engine.util.property.PropertyVector3f;
 import gamelauncher.engine.util.property.PropertyVector4f;
+import org.joml.Math;
 import org.joml.Matrix4f;
-import org.joml.Vector3f;
 import org.joml.Vector4f;
 
 /**
@@ -17,141 +17,133 @@ import org.joml.Vector4f;
  */
 public class GameItem extends AbstractGameResource {
 
-	private final PropertyVector3f position;
-	private final PropertyVector3f scale;
-	private final PropertyVector3f rotation;
-	private final PropertyVector4f color;
-	private final PropertyVector4f addColor;
-	protected Model model;
+    private final PropertyVector3f position;
+    private final PropertyVector3f scale;
+    private final PropertyVector3f rotation;
+    private final PropertyVector4f color;
+    private final PropertyVector4f addColor;
+    protected Model model;
 
-	/**
-	 *
-	 */
-	private GameItem() {
-		position = new PropertyVector3f(0, 0, 0);
-		scale = new PropertyVector3f(1, 1, 1);
-		rotation = new PropertyVector3f(0, 0, 0);
-		color = new PropertyVector4f(1, 1, 1, 1);
-		addColor = new PropertyVector4f(0, 0, 0, 0);
-	}
+    /**
+     *
+     */
+    private GameItem() {
+        position = new PropertyVector3f(0, 0, 0);
+        scale = new PropertyVector3f(1, 1, 1);
+        rotation = new PropertyVector3f(0, 0, 0);
+        color = new PropertyVector4f(1, 1, 1, 1);
+        addColor = new PropertyVector4f(0, 0, 0, 0);
+    }
 
-	public GameItem(Model model) {
-		this();
-		this.model = model;
-	}
+    public GameItem(Model model) {
+        this();
+        this.model = model;
+    }
 
-	public PropertyVector3f position() {
-		return position;
-	}
+    public PropertyVector3f position() {
+        return position;
+    }
 
-	public PropertyVector4f addColor() {
-		return addColor;
-	}
+    public PropertyVector4f addColor() {
+        return addColor;
+    }
 
-	public PropertyVector4f color() {
-		return color;
-	}
+    public PropertyVector4f color() {
+        return color;
+    }
 
-	public PropertyVector3f rotation() {
-		return rotation;
-	}
+    public PropertyVector3f rotation() {
+        return rotation;
+    }
 
-	public PropertyVector3f scale() {
-		return scale;
-	}
+    public PropertyVector3f scale() {
+        return scale;
+    }
 
-	public void addColor(float r, float g, float b, float a) {
-		this.addColor.set(r, g, b, a);
-	}
+    public void addColor(float r, float g, float b, float a) {
+        this.addColor.set(r, g, b, a);
+    }
 
-	public void position(float x, float y, float z) {
-		this.position.set(x, y, z);
-	}
+    public void position(float x, float y, float z) {
+        this.position.set(x, y, z);
+    }
 
-	public void scale(float scale) {
-		this.scale(scale, scale, scale);
-	}
+    public void scale(float scale) {
+        this.scale(scale, scale, scale);
+    }
 
-	public void scale(float x, float y, float z) {
-		this.scale.set(x, y, z);
-	}
+    public void scale(float x, float y, float z) {
+        this.scale.set(x, y, z);
+    }
 
-	public void setColor(float r, float g, float b, float a) {
-		color.set(r, g, b, a);
-	}
+    public void setColor(float r, float g, float b, float a) {
+        color.set(r, g, b, a);
+    }
 
-	public void rotation(float x, float y, float z) {
-		this.rotation.set(x, y, z);
-	}
+    public void rotation(float x, float y, float z) {
+        this.rotation.set(x, y, z);
+    }
 
-	/**
-	 * @return the model
-	 */
-	public Model model() {
-		return model;
-	}
+    /**
+     * @return the model
+     */
+    public Model model() {
+        return model;
+    }
 
-	@Override
-	public void cleanup0() throws GameException {
-		model.cleanup();
-	}
+    @Override public void cleanup0() throws GameException {
+        model.cleanup();
+    }
 
-	/**
-	 * @return the {@link GameItemModel} for this {@link GameItem}
-	 */
-	public GameItemModel createModel() {
-		return new GameItemModel(this);
-	}
+    /**
+     * @return the {@link GameItemModel} for this {@link GameItem}
+     */
+    public GameItemModel createModel() {
+        return new GameItemModel(this);
+    }
 
-	public void applyToTransformationMatrix(Matrix4f transformationMatrix) {
-		transformationMatrix.translate(position.x.floatValue(), position.y.floatValue(),
-				position.z.floatValue());
-		transformationMatrix.rotateX((float) Math.toRadians(-rotation.x.floatValue()));
-		transformationMatrix.rotateY((float) Math.toRadians(-rotation.y.floatValue()));
-		transformationMatrix.rotateZ((float) Math.toRadians(-rotation.z.floatValue()));
-		transformationMatrix.scale(scale.x.floatValue(), scale.y.floatValue(),
-				scale.z.floatValue());
-	}
+    public void applyToTransformationMatrix(Matrix4f transformationMatrix) {
+        transformationMatrix.translate(position.x.floatValue(), position.y.floatValue(), position.z.floatValue());
+        transformationMatrix.rotateX(Math.toRadians(-rotation.x.floatValue()));
+        transformationMatrix.rotateY(Math.toRadians(-rotation.y.floatValue()));
+        transformationMatrix.rotateZ(Math.toRadians(-rotation.z.floatValue()));
+        transformationMatrix.scale(scale.x.floatValue(), scale.y.floatValue(), scale.z.floatValue());
+    }
 
-	/**
-	 * @author DasBabyPixel
-	 */
-	public static final class GameItemModel extends AbstractGameResource
-			implements ColorMultiplierModel, ColorAddModel {
+    /**
+     * @author DasBabyPixel
+     */
+    public static final class GameItemModel extends AbstractGameResource implements ColorMultiplierModel, ColorAddModel {
 
-		/**
-		 *
-		 */
-		public final GameItem gameItem;
+        /**
+         *
+         */
+        public final GameItem gameItem;
 
-		private GameItemModel(GameItem gameItem) {
-			this.gameItem = gameItem;
-		}
+        private GameItemModel(GameItem gameItem) {
+            this.gameItem = gameItem;
+        }
 
-		@Override
-		public void cleanup0() throws GameException {
-			gameItem.cleanup();
-		}
+        @Override public void cleanup0() throws GameException {
+            gameItem.cleanup();
+        }
 
-		@Override
-		public void render(ShaderProgram program) throws GameException {
-			gameItem.model.render(program);
-		}
+        @Override public void render(ShaderProgram program) throws GameException {
+            gameItem.model.render(program);
+        }
 
-		@Override
-		public Vector4f getAddColor() {
-			return gameItem.addColor().toVector4f();
-		}
+        @Override public Vector4f getAddColor() {
+            return gameItem.addColor().toVector4f();
+        }
 
-		@Override
-		public Vector4f getColor() {
-			return gameItem.color().toVector4f();
-		}
+        @Override public Vector4f getColor() {
+            return gameItem.color().toVector4f();
+        }
 
-		public GameItem getGameItem() {
-			return gameItem;
-		}
+        public GameItem getGameItem() {
+            return gameItem;
+        }
 
-	}
+    }
 
 }
